New circle rasterization algorithm

+++ Demos/PhysicsDemo/BallController.cs
@@ -9,7 +9,7 @@ namespace PhysicsDemo
     internal class BallController : BallEntity
     {
 
-        public BallController(Vec2 pos, float radius, float mass) : base(Renderer.GenerateCircleSprite(radius, Color.Salmon), pos, radius, mass) { }
+        public BallController(Vec2 pos, float radius, float mass) : base(Renderer.GenerateCircleSprite((int)radius, Color.Salmon), pos, radius, mass) { }
 
         public override void Update()
         {
